Add first test of tesseract
I needed to see how tesseract would initially react to a stock image I took before committing to it as a library. This is the most base implementation, as the image isn't processed at all. It seems to have worked passably, but there is a lot left to be desired. It has ignored both edges of the receipt (and thus the times bought and prices), and some segments of the bottom receipt, including the total. The formatting is also a bit questionable, so I will have to see what I can do to fix that.